requirements.txt
torch
transformers
accelerate
pypdf
python-docx
python-pptx
PyQt6